Add filter and contextfunction for epoch
Epoch handling for package versions is already supported. But
for handling the epoch field, the plain epoch number is needed.
There is now a new context function called "epoch" (and also
a filter with the same name is added) to handle the epoch.
Example (first one recommended):

Epoch: {{ epoch('oslo.config' }}

or:

Epoch: {{ 'oslo.config' | epoch }}

Both will render to:

Epoch: 2

if there's a epoch.yaml file where the epoch for oslo.config is set to
2.
If the epoch.yaml file is not there, the epoch will be set to 0.

Change-Id: I1aea2bce426b5d1bf105cb177367ce9f74c20a0b