Introduce context variable 'pypi_name'
It will be useful for future functions to know the pypi name inside
of the Jinja2 context. For that a variable must be set.
The first user of that variable is the context function py2name().
If this function is called without any parameter, the value from
'pypi_name' is used for the name calculation.
Also document the currently available context variables.

Change-Id: I92f6d59c287160e8677b2df3ce0a4411a95af760