Add new function fetch_source() to download files
In some cases (eg. tracking stable-$release.tar.gz sdist tarballs) it is
useful to download a given source before rendering the .spec.j2 template.

fetch_source() is doing that and puts the downloaded tarball next to the
output .spec file. The function return the url it takes as parameter so the
usage is like:

{% set source = fetch_source('http://foo/bar.tar.gz') %}
Source0: {{ source }}

Change-Id: Icad8da9094791e2ff67cebe58626b8d76a13eb06