Sort found archives by mtime for version detection
If renderspec finds multiple valid archives to get the version from,
use the archive that was last modified.
The assumption here is, that fetch_source() was called and then the
newly fetched source was last modified and is then used. Otherwise
renderspec might use a previous source to detect the version which
then mismatch with the source to use for the build.

Change-Id: I597e81e27740903c7c61da8056304dda4230410a