Introduce a new --input-template-format switch
This new switch can be used to select a different input template format. The
default will be 'spec.j2' which is the current renderspec behavior.
The plan is to add an extra input format called 'yaml' which can render
a yaml data structure into .spec files for the different spec styles.

Change-Id: Id3ded3ee85f4f32ea6b85d979ca04df97c07cf49