Switch to openstackdocstheme
With the recent Sphinx release the previously used oslosphinx
theme is no longer functional. It is fine to switch over
to the more openstack standard openstackdocstheme which is
maintained and working with the new Sphinx.

Move doc-specific requirements for the docs process into
a separate doc/requirements.py and remove it from test-reqiurements.

Change-Id: I075a02bb9efbfc9c39277bc8ce7330e20bf91605