Preserve list order in _pymod2pkg_translate
Previously, we converted the py_versions list into a set to substract
the versions to be skipped. However, Python does not guarantee the
item order in a set, so some unit tests could fail when the order
was reversed from ['py2', 'py3'] to ['py3', 'py2'].

This change fixes that by making sure the order is maintained.

Change-Id: I5bcdfa3a3dfe2583cc9653d6c87f6c24464b60a5